added demo page and playground
