+++ package.json
@@ -21,6 +21,7 @@
     "@types/katex": "^0.5.0",
     "@types/react-measure": "^2.0.4",
     "@types/react-plotly.js": "^2.2.2",
+    "@types/react-router-dom": "^4.3.1",
     "bootswatch": "git+https://git@github.com/thomaspark/bootswatch.git#95a134434fd0955182e9a938b5e50624e8e143c6",
     "gh-pages": "^2.0.0",
     "i": "^0.3.6",
@@ -46,15 +47,16 @@
     "npm-run-all": "^4.1.3",
     "plotly.js": "^1.40.0",
     "prop-types": "^15.6.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
     "react-measure": "^2.1.2",
     "react-plotly.js": "^2.2.0",
+    "react-router-dom": "^4.3.1",
+    "react-scripts-ts": "^3.1.0",
     "react-test-renderer": "^16.4.2",
     "sanitize-html": "^1.18.4",
     "typescript": "^3.1.6",
-    "uuid": "^3.3.2",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "react-scripts-ts": "^3.1.0"
+    "uuid": "^3.3.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.1.2",
